6827669: CLDC TCK tests failed in StackMap mode
Force CVM_SPLIT_VERIFY=true if CVM_DUAL_STACK=true.

If the split verifier is enabled, then use it for cldc/midp classes, even if no StackMap resource is present.

Fixed bug in split verifier that caused it to assert of the class being verified did not have a StackMap resource.

Reviewed by Dean